Improve type safety, robustness, and performance
- Added compiler optimizations in Dumper.vcxproj for better runtime performance, including advanced instruction sets and parallel code generation.
- Fixed type mismatches using `static_cast` across multiple files to ensure consistent behavior and prevent potential issues.
- Enhanced robustness in Unreal Engine function identification, including improved anchors for `UGameEngine::Tick` and better logic for `StaticConstructObject_Internal`.
- Updated `UEObject::IsA` and `CppGenerator` to handle null `Class` pointers, preventing crashes with malformed objects.
- Improved error handling and type consistency in `NameArray.cpp` and `ObjectArray.cpp`.
- Refined `IsValid` checks to reject invalid or mid-teardown objects, avoiding crashes.
- Improved string handling and range calculations in `PlatformWindows.cpp` for better reliability.
- Enhanced logging and hook tracking in `InstallTickHook` to ensure accurate reporting and avoid redundant operations.

+++ Dumper/Engine/Private/OffsetFinder/OffsetFinder.cpp
@@ -331,7 +331,7 @@ void OffsetFinder::InitFNameSettings()
 	constexpr float MinPercentage = 0.03f;
 
 	/* Minimum required ammount of names for which FName::Number is in a [1...4] range */
-	const int32 FNameNumberThreashold = (ObjectArray::Num() * MinPercentage);
+	const int32 FNameNumberThreashold = static_cast<int32>(ObjectArray::Num() * MinPercentage);
 
 	Off::FName::CompIdx = 0x0;
